Adding :no_cache description to Connected Content
This adds the :no_cache option to Connected Content caching settings. Previously, users were encouraged to get the current date/time and the user ID and add them to the URL. There were some use cases where that wouldn't work (i.e. endpoints that didn't allow for additional GET parameters) and it was hard to explain why we didn't have that option. This adds the option to forego the cache entirely, at the cost of increased send speed.