Update Xamarin SDK Doc with Missing Key Step
This doc was missing a key step (updating the Info.plist file). We spent hours debugging just to figure that out, so hopefully I can save some other people the same headache in the future.

Also, it's really helpful to know how to include verbose logging on iOS for Xamarin, so I included an example of that here as well.