Update intelligent_timing.md
I am a CSM and my customer scheduled an intelligent timing campaign 24hrs before it was due to start. However out of the 50k users he wanted to reach, only 40k were sent the email as this 48hr time period wasn't adhered to the Braze support engineer has told us. 

I feel to make this much more black-and-white as to what is really required for fully campaign delivery the update I have made is required. I'm also happy for you to write this anywhere on the page, OR for it to be in one of the coloured boxes so it really stands out as I think it's an important note to mention.