Update push_reporting.md
Hey Braze team - on this page where it shows the Push analytics UI of Braze for a campaign for opens, sends, clicks, conversation etc... it is showing a CONTENT CARD analytics table and not a Push analytics table. So the analytics available in the screenshot are incorrect as we cannot get all these figures for Push campaigns. 

Can I ask you update this screenshot of the correct Push analytics metrics available?

Thanks,
Andrew