Update changelog.md
Just an idea to help with the customer experience. 
In the Android change log Braze docs, when there is a large update, we have a link under the latest version which takes the customer directly to the GitHub page for that release date: https://www.braze.com/docs/developer_guide/platform_integration_guides/android/changelog/

This isn't on the iOS change log Braze docs page at all, so I wondered how hard is it to add this in so customers can review the data of the SDK releases?