Clarification for Devs Inexperienced with Push
Often when dealing with devs that are inexperienced with push notifications I find that they follow some guide they find online and request push permissions in the startup code. We often need to tell them not to do this. This is more of a problem when the Braze SDK will be handling this for them now.