Update message_step.md
I believe that if the user doesn't meet the delivery settings in Canvas Flow they will exit the Canvas. This is then reflected in the "Exit Canvas" Analytics. 

But a lot of our customers are getting confused bt what "Exited Canvas" means in the analytics so this update to the docs will help.

This is my assumption for how "Delivery settings" work, so please validate this on your side/with Product.