Update link_shortening.md
I believe these docs to be completely wrong when it comes to being able to track "SMS opens". Even further down in the page where it shows the Analytics UI table, there is no column for "opens" only "clicks:", which is what link shortening will allow us to track as Braze owns the link so we'll know when it has been clicked.

This "tracking SMS opens" was also written into the Braze Release Confluence page so it might be worthwhile reviewing that and if we cannot track "opens" to update that Confl page. 

Thanks,
Andrew