Update measuring_and_testing_with_canvas_analytics.md
Hello,

FYI I have just put a comma as an edit in this doc (so I can submit this request) as I wasn't sure where to provide the below info. 


With the new Canvas Flow, when you review Canvas analytics for a "Message Step" there is a new metric title called "Exited Canvas". This is causing a lot of confusion with our customers as there are no docs on what this actually means. 

I believe we need to update this page with the explanation of what this means please as there are screenshots on this page showing the old Canvas analytics UI. 

"There are two ways a user can Exit a Canvas step. 
1. When a user enters the last step in a Canvas Flow, they will immediately exit the Canvas after they have fallen into that step and the message payload has been dispatched This will mean within a few seconds of a user falling into your last Canvas Flow step they will also Exit the Canvas.
2. If "Delivery Settings" has been turned on, if the user does not meet the Audience Segmentation rules set at the step level they will also exit the Canvas after entering the step"