Update exit_criteria.md
Hey team,

Now that this feature is in GA can it please be made a real page in the Braze docs so we can search for it? e.g. on Google or in the Braze Docs search when you search for "Braze Exit Criteria" nothing is returned.

Also - can additional information be added to this page to explain exactly when a customer will fall out of the Canvas based on the exit criteria rules?
e.g. if there is a delay step of 5 days that the customer is currently being delayed in, does the user fall out of the Canvas instantly when the Event or Segment referenced in the Exit Criteria is met? OR do they fall out of the Canvas after the 5 day delay has been processed and Braze determines who needs to progress through the Canvas?

Nuances like the above to cover in the docs will help our customers understand the functionality much better and reduce support cases raised.

Thanks :)