Reword IP Range to IPs for Webhook IP Whitelist
While I was on call there was some confusion about the fact. we talk about the Whitelist IPs for webhook as ranges. This caused confusion for a client.

Adam Kosmin 5 hours ago
I reviewed the list and have to believe the documentation on that page is a bit misleading since the word "range" is used. Given the fact so many of the listed IPs are clearly in the same network, any CIDR notation would be /32. I think that page should have the word "range" removed.