Update user_permissions.md
To be able to edit a customers email or push subscription status (subscribed vs opted-in) and their Push/email subscription groups, this "PII" dashboard user setting needs to be enabled. But no where in the docs does it say this. 

This has confused a few customers of ours recently so we're hoping the docs can be updated to reflect the exact functionality of this toggle.

Thanks,
Andrew