Updating interactions list
When updating this page to include a new interaction `Clicked shortened SMS link`, I noticed that the rest of the list is out of date with what is available in the platform. Thus, I went through and added all the missing interactions by comparing the documentation to what is seen in Sweeney. It would be good to have the Docs team review that for accuracy, as I initially intended to just add 'Clicked shortened SMS link' as that is the only new interaction that I own, but figured I would make the additional updates while I was in this file.