Update preference_center_via_api.md
Hello,

We've had a number of customers ask how the Braze API preference centre works, and to be honest, we as Braze CSMs have also found the documentation to be extremely confusing. 

Based on learning now how the API custom Braze Preference centre now works, can I recommend there is some verbiage at the very top of the main page that contains copy similar to "The Braze API Preference Centre is a Braze hosted website that can display your customer's subscription state and subscription group statuses. Using HTML and CSS, your engineering team can build your Preference Centre from scratch, however you like, using HTML and CSS so that the styling of the page matches your brand guidelines. Utilising Liquid also enables you to retrieve the names of your Subscription Groups, and also the individual user's status. This way Braze stores and retrieves this data when the page is loaded."

The above explanation now tells customers that the preference centre is actually a completely blank website that they can code however they want. And that Braze is hosting the page for them. Currently, we've had a lot of customers assume the out of the box Braze preference centre is the starting blocks, and that they need the HTML tags for h1, h2, h3, fonts, the radio buttons etc so that the HTML they provide aligns to what Braze has already developed. 

We'd like this feedback to be strongly considered as it's increasing the number of support cases, and confusion amongst customers. 

Thanks :)