Update managing_user_subscriptions.md
DId you guys delete the "update preferences" Liquid code from this page as it used to be here...

E.g. in the footer a user can "Click here to unsubscribe" which Brae provides that Liquid. But you also used to say "Click here to update your preference centre" which included the new Liquid I've put back in...

I assume this still works as I have customers actively using this still...