Update nested_custom_attribute_support.md
In conversation with Kevin (Braze), he mentioned: "I have received confirmation from the product team that setting a nested custom attribute via SDK is not currently available. However, this is on the roadmap to implement currently. As a workaround I would suggest setting the nested custom attribute via API using the users track endpoint."

We've found this issue while implementing a solution for Hayu.com.