Update snowflake.md
A number of customers in Australia are getting confused when referencing the Snowflake data schema as the file naming convention differs from everywhere else in Braze. 

e.g braze_id and external_id/user_id are now labelled something different in Snowflake, so we felt this should 100% be highlighted somewhere. (e.g. my update to this Docs page)