Update naming_convention.md
A few customers here in Australia have flagged that the Snowflake data schema references "user_id" which is actually the "braze_id" and "external_user_id" which is actually the "external_id". And said it's very confusing as no where does it explain what the "new" fields in the Snowflake data schema mean/map to. 

The Snowflake data schema page is being updated to explain the difference between the Snowflake data schema and the Braze data schema. And one of the customers who complained also highlighted this page as not even referencing the "external_id". 

Perhaps Braze is transitioning away from "external_id" however I wanted to propose this edit so that it's obvious to all customers that "external_user_id" also means "external_id"...