Update ssl.md
I am nearly 100% sure that this step "Reach out to your customer success manager to get SSL turned on." in setting up an SSL cert is incorrect. 

Once the customer has setup their SSL certificate themselves, pointing the DNS to their CDN, they are good to go so I feel this step should be removed. 

Maybe check with Email Ops, but I am nearly 100% confident in this as I've never had to validate a customers SSL setup once it's locked in.