Update html_in-app_messages.md
We've had a few customers of late trying to be very technical with HTML IAMs and JS Bridge. 

They have wanted to retreive a saved customAttribute stored on a users profile into their JS Bridge method (e.g. a button click) then update that customAttribute when the button is clicked. 

But without knowing the current user's customAttribute value, they cannot define the required needs in the JS Bridge call. 

So we wondered does this extra line of text on the Braze docs helps clarify Liquid cannot be used within JS Bridge?