Update email_capture_form.md
I am quite confident that if you send a user profile a new email address (one that differs from the currently saved email) and the original email = hard bounce. Then the hard bounce status is removed (as the new email hasn't hard bounced yet). 

Thought this section would be a good place to list this as I thought it was listed elsewhere in the docs but cannot see it anywhere.