Update faq.md
Hello - this error message appears I believe for any customer when they are creating an email in Braze... but we don't link them to any docs which explain why this error message is appearing. 

I believe my explanation in this update covers why the error message appears... but it might be worth validating this to 100% confirm. 

Otherwise, I think we need this information in the FAQ to provide better clarity as a lot of customers use Liquid to insert an unsubscribe URL and the error still appears and customers then submit support cases on this.