Update purchase_events.md
We are very worried about the calculation here: 

$$\text{Lifetime revenue} = \frac{\text{Total spend in dollars}}{\text{Total number of purchase events}}$$  

For us, this should be Lifetime Revenue PER USER as the 2nd bullet point below this calculation is directly relating to this calculation and nothing else. 

I've had 3 Snr CSMs reviewing this for an hour to try and understand this and we've just realised the above needs to be resolved.