Update revenue_report.md
We think there is a huge flaw in our docs here where we don't explain at all the calculations for the top 3 graphs on the Revenue Report page. 

We've had 3 x Snr CSMs spend 1hr scouring the docs and trying to figure out where on earth all these metrics are coming from and if they are unique and de-duplicated outside of the campaigns and Canvases. 

We think the docs has to be updated to have this information we've inserted as it's now crystal clear (happy for you to fix our proposal copy up) what each of those metrics is.