Update authentication.md
I would like to recommend the instructions to check DMARC "Option 1: You can input your parent domain or subdomain in any third-party DMARC checker, such as MXToolbox , to audit whether you have a DMARC policy in place and what that policy is set to."

Are edited to be much more specific. 

From my experience, you cannot "just check the DMARC for the domain" in MX Tools. You need to specifically list the parent domain in MX Tools vs the sub-domain the Braze customer is emailing from. 
E.g. MX Tools cannot look at the Braze sub-domain to see if the DMARC is present as it will 98% of the time have been defined at the root domain. And MX Tools doesn't look "upwards" for parent DMARC certs. 

This means customers get an error and say the Braze docs are incorrect and that they need to set the DMARC at the child/subdomain. 

So I think much better detail needs to be provided in these instructions. e.g. 

If you have set your DMARC are the root domain, insert that into MX Tools. if you have set the DMARC at the sub-domain, enter the sub-domain into MX Tools. (FYI, MX Tools doesn't "look up/down" when doing it's lookups. So if you have set the DMARC at the root domain and enter the sub-domain, MX Tools will show a failure as it doesn't know the DMARC has been set at the root domain"

Something like the above will avoid a lot of confusion by customers