Update using_locales.md
So in the instructions, it doesn't actually very clearly explain you need to add "id_1", "id_2" etc to the email to include your translations. 

ALSO - we've found a bit of an issue with a few of our customers. To remove human error/fat fingers we also say "copy the code/liquid" from the Braze website. 

As the "{ % translations % }" test in the instructions is like coded as "code" (e.g. it is in a code font and a grey background), when the user copies this direct from the Braze website and pastes it into the DnD email editor... when you download the Language translation CSV from Braze, it includes "<code>" tags in the language file which breaks everything. 

Example here is my Loom video: https://www.loom.com/share/204954bed3a34b359c9a6d8516b1e610?sid=2c181c15-71f6-45a6-9612-6650703f3118

THE ASK: is there anyway to remove the "code" format against the translation Liquid in the docs to remove huge confusion for our customers?