Update security_settings.md
Under the "## Downloading a security event report" heading, are we able to list out a more extensive list of what exactly is able to be downloaded from this report? Often customers ask, but we have no long list of event types that are include in this report... so updating the docs would be very helpful please.