Update engagement_reports.md
Is there anyway Braze can be super clear on the date range/data.

e.g. when I select the date range to create the report, is this the date range based on the Dashboard Company Timezone? or Based on Braze's UTC timezone?

It doesn't list this anywhere clearly in Braze