Update dynamics_365_customer_insights.md
I removed the first line of Step 2 which seemed to be an erroneous repeat of the first line from Step 1. This essentially started the step which explained how segments should be set up in braze with instructions about how to access a page in the dynamic 365 platform. 

The customer flagged this as confusing, and suggested we put step 2 in front of step 1, but having reviewed the documentation, I found that after removing the mistake it made more sense to leave step 2 where it was.