Update data_retention.md
I cannot figure out how to get to a secondary page sorry, however the User Archival dedicated page here: https://www.braze.com/docs/user_archival

Can we PLEASE add copy at the start of the page which clearly calls out that Braze "fully deletes all dormant users from our system"?

We've now had a 2nd customer in the last month here in Australia threaten legal action as they understood "user archival" to mean "the user is archived and can be brought back to life". Vs Brazes definition of "archival" as "we fully delete the user as they cannot be brought back to life".

It's a small nuance in terminology, but super critical to making sure customers fully understand the reality this Braze rule can cause if they are a bit aloof with it's severity.

Thanks for taking this into consideration as from the ANZ team this is critical to call out the reality that users a completely deleted from the Braze system.